+++ src/tests/config.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "@jest/globals";
 import request from "supertest";
-import { app } from "../index";
+import { app } from "../app";
 
 describe("Test config", () => {
   it("returns 200 and has config", async () => {
